refactor: auction history schedule 로직 별도의 클래스로 분리
