+++ src/main/java/until/the/eternity/auctionhistory/domain/dto/external/OpenApiAuctionHistoryResponse.java
@@ -1,8 +1,9 @@
 package until.the.eternity.auctionhistory.domain.dto.external;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import java.util.List;
 import until.the.eternity.itemoption.domain.dto.external.OpenApiItemOptionResponse;
+
+import java.util.List;
 
 public record OpenApiAuctionHistoryResponse(
         @JsonProperty("item_name") String itemName,
