+++ src/main/java/until/the/eternity/auctionhistory/domain/mapper/OpenApiAuctionHistoryAssembler.java
@@ -1,12 +1,13 @@
 package until.the.eternity.auctionhistory.domain.mapper;
 
-import java.time.OffsetDateTime;
-import java.util.List;
 import org.springframework.stereotype.Component;
 import until.the.eternity.auctionhistory.domain.dto.external.OpenApiAuctionHistoryResponse;
 import until.the.eternity.auctionhistory.domain.dto.internal.response.AuctionHistoryDetailResponse;
 import until.the.eternity.auctionhistory.domain.dto.internal.response.ItemOptionResponse;
 import until.the.eternity.common.enums.ItemCategory;
+
+import java.time.OffsetDateTime;
+import java.util.List;
 
 @Component
 public class OpenApiAuctionHistoryAssembler {
