+++ src/main/java/until/the/eternity/auctionhistory/repository/AuctionHistoryRepository.java
@@ -1,7 +1,5 @@
 package until.the.eternity.auctionhistory.repository;
 
-import java.util.List;
-import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
@@ -11,6 +9,9 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import until.the.eternity.auctionhistory.domain.entity.AuctionHistory;
+
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface AuctionHistoryRepository
