+++ src/main/java/until/the/eternity/auctionhistory/repository/AuctionHistoryRepositoryImpl.java
@@ -2,7 +2,6 @@ package until.the.eternity.auctionhistory.repository;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
@@ -10,6 +9,8 @@ import org.springframework.data.domain.Pageable;
 import until.the.eternity.auctionhistory.domain.dto.internal.request.AuctionHistorySearchRequest;
 import until.the.eternity.auctionhistory.domain.entity.AuctionHistory;
 import until.the.eternity.auctionhistory.domain.entity.QAuctionHistory;
+
+import java.util.List;
 
 @RequiredArgsConstructor
 public class AuctionHistoryRepositoryImpl implements AuctionHistoryRepositoryCustom {
