+++ src/main/java/until/the/eternity/auctionitem/domain/entity/AuctionItem.java
@@ -1,10 +1,11 @@
 package until.the.eternity.auctionitem.domain.entity;
 
 import jakarta.persistence.*;
-import java.time.LocalDateTime;
-import java.util.List;
 import lombok.*;
 import until.the.eternity.itemoption.domain.entity.ItemOption;
+
+import java.time.LocalDateTime;
+import java.util.List;
 
 @Entity
 @Table(name = "auction_item")
