+++ src/main/java/until/the/eternity/itemminprice/service/ItemDailyMinPriceService.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import until.the.eternity.itemminprice.repository.ItemDailyMinPriceRepository;
 
-/** 일간 최저가 upsert 배치를 실행하는 서비스. */
 @Slf4j
 @Service
 @RequiredArgsConstructor
