+++ src/main/java/until/the/eternity/itemoption/domain/entity/ItemOption.java
@@ -1,7 +1,10 @@
 package until.the.eternity.itemoption.domain.entity;
 
 import jakarta.persistence.*;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 import until.the.eternity.auctionhistory.domain.entity.AuctionHistory;
 import until.the.eternity.auctionitem.domain.entity.AuctionItem;
